fix(runner): duplicate `y` at end of spinner
This has got to be the most stupid bug I have encountered. It turns out
that the spinner uses the `\r` character to clear the existing spinner or
something, which doesn't really work? Seems like a hack. Anyway, this
resulted in the previous message not clearing completely. The
workaround for this issue is to add an extra space after the main
message to overwrite the underlying spinner message.